+++ port/esp_idf/components/golioth_sdk/CMakeLists.txt
@@ -26,6 +26,7 @@ idf_component_register(
     SRCS
         "${sdk_port}/freertos/golioth_sys_freertos.c"
         "${sdk_port}/esp_idf/fw_update_esp_idf.c"
+        "${sdk_port}/esp_idf/golioth_sys_espidf.c"
         "${sdk_port}/utils/hex.c"
         "${sdk_src}/golioth_status.c"
         "${sdk_src}/coap_client.c"
