+++ port/modus_toolbox/golioth_port_config.h
@@ -5,14 +5,6 @@
  */
 #pragma once
 
-/* ModusToolbox uses an older 2.25 version of mbedtls
- * These function names changed in version 3.0:
- * https://github.com/Mbed-TLS/mbedtls/commit/26371e47939d02d5f1a5ad9012f49aff0c7f5bc4
- */
-#define mbedtls_sha256_starts mbedtls_sha256_starts_ret
-#define mbedtls_sha256_update mbedtls_sha256_update_ret
-#define mbedtls_sha256_finish mbedtls_sha256_finish_ret
-
 // Same as GLTH_LOGX, but with 32-bit timestamp (PRIu32),
 // since MTB built-in C lib doesn't support 64-bit formatters.
 #define GLTH_LOGX(COLOR, LEVEL, LEVEL_STR, TAG, ...) \
